restructure migrations
