+++ public/app/plugins/datasource/cloudwatch/components/usePreparedMetricsQuery.ts
@@ -1,6 +1,7 @@
 import deepEqual from 'fast-deep-equal';
 import { useEffect, useMemo } from 'react';
 
+import { migrateMetricQuery } from '../migrations/metricQueryMigrations';
 import { CloudWatchMetricsQuery, MetricEditorMode, MetricQueryType } from '../types';
 
 export const DEFAULT_QUERY: Omit<CloudWatchMetricsQuery, 'refId'> = {
@@ -21,10 +22,11 @@ export const DEFAULT_QUERY: Omit<CloudWatchMetricsQuery, 'refId'> = {
 
 const prepareQuery = (query: CloudWatchMetricsQuery) => {
   const withDefaults = { ...DEFAULT_QUERY, ...query };
+  const migratedQuery = migrateMetricQuery(withDefaults);
 
   // If we didn't make any changes to the object, then return the original object to keep the
   // identity the same, and not trigger any other useEffects or anything.
-  return deepEqual(withDefaults, query) ? query : withDefaults;
+  return deepEqual(migratedQuery, query) ? query : migratedQuery;
 };
 
 /**
